Generalised KineticsReaderFactory for use elsewhere
